+++ Sources/KeyboardKit/Actions/SecondaryCalloutActionProvider.swift
@@ -3,6 +3,7 @@
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-01-06.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Foundation